Rename helpers.js -> colors.js
